// backups/2026-03-10-wishes-fixes/types.ts
export type ActivityLevel = 'NORMAL' | 'SESSION' | 'WHITE_GREEN' | 'REDUCED' | 'CLOSED';

export type RecalcScope = 'none' | 'day' | 'week' | 'rest-month';

export type WorkZone = 
  | 'URGENCES' 
  | 'TRAVAIL' 
  | 'URGENCES_TARDE'
  | 'TRAVAIL_TARDE'
  | 'ADMIN' 
  | 'ADM_PLUS'
  | 'TW' 
  | 'TW_ABROAD'
  | 'F' 
  | 'STRASBOURG' 
  | 'LIBERO'
  | 'RECUP'
  | 'FP'
  | 'CS'
  | 'SICK_LEAVE'
  | 'CA'
  | 'VACCIN'
  | 'VACCIN_AM'
  | 'VACCIN_PM'
  | 'VACCIN_PM_PLUS'
  | 'TRAVAIL_C'  // Complemento clínico corto
  | 'URGENCES_C'; // Complemento clínico corto

export interface Shift {
  id: WorkZone;
  label: string;
  description: string;
  color: string;
  textColor: string;
}

export interface Nurse {
  id: string;
  name: string;
  email: string;
  role: UserRole;
  password?: string;
  order: number;
  mustChangePassword?: boolean;
  passwordResetRequired?: boolean;
}

export type UserRole = 'admin' | 'nurse' | 'viewer';
// ...
export type Agenda = {
  [weekIdentifier: string]: ActivityLevel; // e.g., '2024-W34'
};
// ...
export interface Wish {
  text: string;
  validated: boolean;
  shiftType?: WorkZone;
}

export interface Wishes {
  [nurseId: string]: {
    [dateKey: string]: Wish;
  };
}